RESKC-608: Fixing role member fetch. "Proposal Creator" role is behaving like a Unit Hierarchy role with descend enabled in stead of Unit role sometimes. User with the "Unit" Proposal Creator role assigned at a unit that has children, has all the children listed in the PD "Lead Unit" drop down when Creating or Copying a proposal, even if they do not have the role assigned for the child units. User can select, create/copy and successfully save the new proposal.
After testing I think this behavior is random. I assigned coiadmin to the Proposal Creator role with unit BL-BL and had access to ALL units not just its descendents. The behavior seems to be dependent on the other roles assigned to the user. This is because of how rice fetches role members when given a role id and user id.